Splitting up ShareTimetableModal into separate components under /share directory
